use math multiplyexact for multiplication (#1095)
* Use Math.multiplyExact to guard against integer overflow in multiplications

* int exact

* avoid some casts

* Update Ole10Native.java

* fix test issues

---------

+++ poi-scratchpad/src/main/java/org/apache/poi/hemf/record/emf/HemfHeader.java
@@ -173,7 +173,7 @@ public class HemfHeader implements HemfRecord {
             if (skip < 0 || descriptionEnd > recordSize + HEADER_SIZE || skip + descriptionBytes > Integer.MAX_VALUE) {
                 throw new RecordFormatException("Invalid EMF header description bounds");
             }
-            int maxDescriptionLength = (int)Math.min(recordSize, Integer.MAX_VALUE);
+            int maxDescriptionLength = Math.toIntExact(Math.min(recordSize, Integer.MAX_VALUE));
             IOUtils.safelyAllocateCheck(descriptionBytes, maxDescriptionLength);
             leis.mark((int)(skip + descriptionBytes));
             leis.skipFully((int)skip);
